Fix android build problem after .35 upgrade
