+++ example/android/app/src/main/AndroidManifest.xml
@@ -14,7 +14,8 @@
       android:allowBackup="true"
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
-      android:theme="@style/AppTheme">
+      android:theme="@style/AppTheme"
+      android:name=".MainApplication">
       <activity
         android:name=".MainActivity"
         android:label="@string/app_name"
